fix bug#959 【委员弹劾】cr_02委员弹劾成功后，进入界面加载后报错，其他委员详情界面暂无问题

+++ UnionSquare.Android/app/src/main/java/org/elastos/wallet/ela/ui/committee/bean/CtDetailBean.java
@@ -36,7 +36,7 @@ public class CtDetailBean extends BaseEntity {
         private long endDate;
         private double impeachmentThroughVotes;
         private double impeachmentVotes;
-        private long impeachmentRatio;
+        private double impeachmentRatio;
         private String depositAmount;
         private List<Term> term;
 
@@ -192,7 +192,7 @@ public class CtDetailBean extends BaseEntity {
             this.impeachmentVotes = impeachmentVotes;
         }
 
-        public long getImpeachmentRatio() {
+        public double getImpeachmentRatio() {
             return impeachmentRatio;
         }
